fix: improve error in type checker
The error from the type checker when attempting to retrieve a trait type
was incorrect. If the current contract is "foo", and the type "bar.baz"
is not found in the analysis database, the error message said:

```
use of unresolved contract 'ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM.foo'
```

When the actual problem was that the contract "bar" was not found.

Similarly, if the contract "bar" is found, but the trait "baz" is not
found, the same error was given, instead of correctly showing that the
trait "baz" was not found.

See #3064