feat: add placeholder type and refactor lexer
The placeholder type is used to hold the original string from the source
when the code is invalid and the lexer/parser reports an error.

The lexer has been refactored such that every token indicates the exact
string from the source file. The later processing of the string happens
in the parser. Only invalid character type of errors should be reported
by the lexer.